+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Fixed
+- When use `changeKeys` param and cannot get any items, it will return an empty object (before returns an empty array)
+
 ## [5.6.3] - 2021-08-09
 ### Fixed
 - Fixed DB connections cache for configs from AWS Secrets Manager